WIP: Fix type casting bugs in ActiveRecord::Relation#pluck
The bug #32856 is caused because after executing the select statement,
the result relationship losses the table references. For instance, if
our query was `authors.id`, it becomes just `id`. Then, when we cast
the values, for the `id` case ActiveRecord will look for the type for
that column in `klass.cast_values` and will use the wrong type.