Use default session options if a previous session comes without having set it
Rails overrides the Rack's request `session` and `session_options` by
defining its own initializers and accessors.

However, in the case of `session_options`, it could happen that if the
session is accessed in a middleware (eg. `DatabaseSelector`) previous
to a store one without having tried to access the session options (and
by this initializing them), it will fail to write the session id in
`ActionDispatch::Request::Session#load?`.

This is a different approach than the used in #42216, that could be a
bit more in line to how Rails initializes these two objects.